+++ boot_junion/src/main/java/com/boot/Controller/ComNoticeController.java
@@ -36,13 +36,16 @@ import org.springframework.web.bind.annotation.ResponseBody;
 import org.springframework.web.multipart.MultipartFile;
 
 import com.boot.DAO.ComNoticeDAO;
+import com.boot.DTO.CardPageDTO;
 import com.boot.DTO.ComNoticeAttachDTO;
 import com.boot.DTO.ComNoticeDTO;
 import com.boot.DTO.ComScrapDTO;
 import com.boot.DTO.RecentNoticeDTO;
 import com.boot.DTO.ResumeDTO;
+import com.boot.DTO.Standard;
 import com.boot.DTO.SubmitDTO;
 import com.boot.DTO.UserDTO;
+import com.boot.Service.CardPageService;
 import com.boot.Service.ComNoticeService;
 import com.boot.Service.IndividualService;
 import com.boot.Service.ScrapService;
@@ -60,20 +63,87 @@ public class ComNoticeController {
 	@Autowired
 	private ScrapService scrapService;
 	
+	@Autowired
+	private CardPageService pageService;
+	
 
 	@RequestMapping("/jobPostList")
 //	public String jobPost(Model model, HttpSession session) {//대메뉴 -> 채용 클릭시 이동(페이징 처리를 위해 CardPageController 매핑됨)
-	public String jobPost(Model model) {//대메뉴 -> 채용 클릭시 이동(페이징 처리를 위해 CardPageController 매핑됨)
-		log.info("jobPost");
+//	public String jobPost(Model model) {//대메뉴 -> 채용 클릭시 이동(페이징 처리를 위해 CardPageController 매핑됨)
+		public String jobPost(Standard std,Model model, HttpSession session) {//대메뉴 -> 채용 클릭시 이동(페이징 처리를 위해 CardPageController 매핑됨)
+//		log.info("jobPost");
+//		
+//		
+//		ArrayList<ComNoticeDTO> dto = postService.JobPostCard();
+//		model.addAttribute("jobPost", dto);
 		
 		
-		ArrayList<ComNoticeDTO> dto = postService.JobPostCard();
-		model.addAttribute("jobPost", dto);
+		log.info("@# cardPage controller");
+		log.info("@# cardPage controller std!!=>"+std);
+		
+		ArrayList<ComNoticeDAO> list = pageService.cardPageList(std);//현재 진행중인 공고를 가져옴
+		int total = pageService.getTotalCount();
+		log.info("총 게시글 수는??"+total);
+		
+		model.addAttribute("jobPost", list);
+		model.addAttribute("paging", new CardPageDTO(total, std));
 
+		
+		 String user_email = (String) session.getAttribute("login_email");//세션에 저장된 사용자이메일 가져오기
+		 log.info("@# jobPost user_email => "+user_email);
+		 
+		 if(user_email != null) {// 세션에 이메일 값이 있다면 해당 사용자의 이메일을 기반으로 관심 공고 목록을 가져옴
+			 ArrayList<Integer> noticeList = scrapService.getScrapNoticeNum(user_email);
+			 
+			 model.addAttribute("noticeList", noticeList);
+			 }
+		
 		
 		return "/recruitmentNotice/jobPostList";
 	}
 
+	
+	// 2024-08-01 지수 (공고 목록 사진 들고오기)
+//		이미지파일을 받아서 화면에 출력(byte 배열타입)
+		@GetMapping("/cardPageDisplay")
+		public ResponseEntity<byte[]> getFileAttach(String fileName) {
+			log.info("@# display fileName=>"+fileName);
+			
+//			업로드 파일경로+이름
+			File file = new File("C:\\devv\\upload\\"+fileName);
+			log.info("@# file=>"+file);
+			
+			ResponseEntity<byte[]> result=null;
+			HttpHeaders headers=new HttpHeaders();
+			
+			try {
+//				파일타입을 헤더에 추가
+				headers.add("Content-Type", Files.probeContentType(file.toPath()));
+//				파일정보를 byte 배열로 복사+헤더정보+http상태 정상을 결과에 저장
+				result = new ResponseEntity<>(FileCopyUtils.copyToByteArray(file), headers, HttpStatus.OK);
+			} catch (Exception e) {
+				e.printStackTrace();
+			}
+			
+			return result;
+		}
+		
+		@GetMapping(value = "/cardPageFileList")
+		public ResponseEntity<List<ComNoticeAttachDTO>> cardPageFileList(@RequestParam HashMap<String, String> param) {
+			log.info("@# cardPageFileList()");
+			log.info("@# param=>"+param);
+			log.info("@# param=>"+param.get("notice_num"));
+			
+			return new ResponseEntity<>(pageService.cardPageFileList(Integer.parseInt(param.get("notice_num"))), HttpStatus.OK);
+		}
+
+	
+	
+	
+	
+	
+	
+	
 	
 	@RequestMapping("/jobPostDetail")
 	public String JobPost(HttpSession session, int notice_num, Model model) {//채용공고 목록 -> 채용공고 상세 이동
