+++ boot_junion/src/main/java/com/boot/Service/CardPageService.java
@@ -10,6 +10,7 @@ import com.boot.DTO.Standard;
 public interface CardPageService {
 	public ArrayList<ComNoticeDAO> cardPageList(Standard std);// 페이징을 위한 메소드
 //	public ArrayList<ComNoticeDAO> cardPageList(CardPageDAO dao);// 페이징을 위한 메소드
+	public int getTotalCount();//전체 공고 수 구하기
 	
 	public List<ComNoticeAttachDTO> cardPageFileList(int notice_num); // 2024-08-01 지수 (공고 목록 사진 들고오기)
 }
